requirements.txt
bert-score==0.3.13 
cohere==5.3.4
networkx==3.3
nltk==3.8.1 
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
rouge==1.0.1
scikit-learn==1.4.2
transformers==4.40.1
spacy==3.7.3